refactor: Endpoint package moved into rules package
