+++ internal/rules/endpoint/basic_auth_strategy.go
