+++ internal/rules/endpoint/basic_auth_strategy_test.go
