+++ internal/rules/endpoint/endpoint.go
