+++ internal/rules/endpoint/endpoint_test.go
