+++ internal/rules/endpoint/mapstructure_decoder_test.go
