+++ internal/rules/endpoint/noop_renderer.go
