+++ internal/rules/endpoint/renderer.go
