+++ internal/rules/mechanisms/contextualizers/config_decoder.go
@@ -19,7 +19,7 @@ package contextualizers
 import (
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/dadrus/heimdall/internal/endpoint"
+	"github.com/dadrus/heimdall/internal/rules/endpoint"
 	"github.com/dadrus/heimdall/internal/rules/mechanisms/template"
 )
