+++ internal/rules/provider/httpendpoint/ruleset_endpoint.go
@@ -26,9 +26,9 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/dadrus/heimdall/internal/endpoint"
 	"github.com/dadrus/heimdall/internal/heimdall"
 	"github.com/dadrus/heimdall/internal/rules/config"
+	"github.com/dadrus/heimdall/internal/rules/endpoint"
 	"github.com/dadrus/heimdall/internal/x/errorchain"
 )
